+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Trust/TrustContacts.cs
@@ -1,0 +1,9 @@
+﻿namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Trust;
+
+public record TrustContacts(
+    Person? TrustRelationshipManager,
+    Person? SfsoLead,
+    Person? AccountingOfficer,
+    Person? ChairOfTrustees,
+    Person? ChiefFinancialOfficer
+);